redpill v1.2.6 - SilenceDefender using firewall rules
Windows Defender for Endpoint (formerly Windows Defender ATP) is a so-called “cloud powered” EDR product,
i.e. alerts and events are pushed to the cloud where defenders can respond to them. When doing Red Team
assignments one of the biggest hurdles usually lie in evading EDR products and ensuring that our actions
are not detected. While a lot of work and research has been put into evading and bypassing Windows Defender
for Endpoint, little research explores the possibility of simply silencing MD for Endpoint such that no data
is sent to the cloud.